refactor: Update Analytics page with GDPChart component
This commit updates the Analytics page by replacing the AreaChart component with the GDPChart component. The AreaChart component was commented out and the GDPChart component from '@/components/gdpWrapper' was imported and used instead. Additionally, a sample data array was added to the page to test the functionality of the GDPChart component. This change improves the visualization of the Bitcoin Price data on the Analytics page.